Extract cmdutil package
